% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{addDDict}
\alias{addDDict}
\title{Add a Variable to a Data Dictionnary}
\usage{
addDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{vars}{Named vector of variables to add to the DDictionary.}
}}
}
\value{
Object of class \code{DDict}.
}
\description{
Add a Variable to a data Dictionary.
}
\details{
Add information about a variable to a DDictionary. The vector \code{vars}
must include the same information, in any order, as the one found  in the
\code{data} property of \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
