% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{castDDict}
\alias{castDDict}
\title{Cast Columns' Data Type Using a \code{DDict}.}
\usage{
castDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame with variabe=les to rename.}
\item{table_nm}{Name of the table.}
}}
}
\value{
\code{data} with modified data types.
}
\description{
Cast columns' data type using a \code{DDict}.
}
\details{
The information is stored in an object of class \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
