% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols.R
\name{clean_cols_constant}
\alias{clean_cols_constant}
\title{Find Variables With a constant Value}
\usage{
clean_cols_constant(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Integer() with positions and names.
}
\description{
Find Variables with a constant value.
}
\details{
A named integer() will be returned with
the names and positions of the columns having a constant value.
}
\examples{
\dontrun{
TODO
}
}
