% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols.R
\name{clean_cols_uniq}
\alias{clean_cols_uniq}
\title{Find Variables With Unique Value}
\usage{
clean_cols_uniq(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Integer() with positions and names.
}
\description{
Find Variables with unique value.
}
\details{
Find Variables with unique value. A named integer() will be returned with
the names and positions of the columns having a unique value.
}
\examples{
\dontrun{
TODO
}
}
