% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dat.R
\name{clean_dat_oob}
\alias{clean_dat_oob}
\title{Replace out-of-bound date with \code{NA}}
\usage{
clean_dat_oob(data, start = "2000-01-01", end = "2050-01-01")
}
\arguments{
\item{data}{Data frame.}

\item{start}{String with start date. Converted to \code{POSIXct} internally.}

\item{end}{String with end date. Converted to \code{POSIXct} internally.}
}
\value{
Data frame with modified \code{POSIXct} columns.
}
\description{
Replace out-of-bound date with \code{NA}.
}
\details{
Go through every column in \code{data} that are of type \code{POSIXct} only.
The interval between \code{start} and \code{end} is tested as
\code{(x >= start) & (x < end)}.
}
\examples{
n <- 9L
df <- data.frame(
  ndx = seq_len(n),
  int = rpois(n, lambda = 5),
  dat = sample(
    c(as.POSIXct(Sys.Date()), as.POSIXct("1900-01-01")),
    size = n, replace = TRUE
  )
)
out <- clean_dat_oob(df, start = "2000-01-01", end = "2050-01-01")
out
}
