% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_empty}
\alias{clean_empty}
\title{Replace empty string with \code{NA}}
\usage{
clean_empty(
  data,
  na_patterns = c("^[[:punct:]]$", "_NA", "^[Nn][/\\\\\\\\[[:space:]]]*[Aa]$")
)
}
\arguments{
\item{data}{Data frame.}

\item{na_patterns}{Character vector of regular expressions to run with
\code{stringr::str_replace(x, pattern = stringr::regex(pattern, ignore_case = TRUE),
replacement = NA_character_)}.}
}
\value{
Data frame.
}
\description{
Replace empty string with \code{NA}.
}
\details{
Replace empty string with \code{NA}. The extra white spaces are removed with
\code{stringr::str_squish}. Important: This is done before the
conversions to \code{NA}. So that "  " will actually be treated as "".
. This behavior affects all the strings.
}
\examples{
\dontrun{
TODO
}
}
