% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_inf}
\alias{clean_inf}
\title{Replace \code{NaN, -Inf, Inf} with \code{NA}}
\usage{
clean_inf(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame.
}
\description{
Replace \code{NaN, -Inf, Inf} with \code{NA}.
}
\details{
The \code{NaN} don't need to be replaced as they are considered to be
\code{NA}. That is \code{is.na(x)} is \code{TRUE} when \code{x = NaN}.
}
\examples{
\dontrun{
TODO
}
}
