% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_NA}
\alias{clean_NA}
\title{Replace empty string and \code{NaN, -Inf, Inf} with \code{NA}}
\usage{
clean_NA(
  data,
  na_patterns = c("^[[:punct:]]$", "^[Nn][/\\\\\\\\][Aa]$"),
  na_strings = c("", "_NA")
)
}
\arguments{
\item{data}{Data frame.}

\item{na_patterns}{Character vector of regular expressions to run with
\code{stringr::str_replace(x, pattern = pattern, replacement = NA_character_)}.
Default value is \code{c("^[[:punct:]]$", r"(^[Nn][/\\][Aa]$)")}.}

\item{na_strings}{Character() with strings to b replaced by \code{NA}. You
must always keep "n/a" and "".}
}
\value{
Data frame.
}
\description{
Replace empty string and \code{NaN, -Inf, Inf} with \code{NA}.
}
\details{
This function calls \code{clean_nan_inf} and \code{clean_empty}.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
clean_inf clean_empty
}
