% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CScaler.R
\name{CScaler}
\alias{CScaler}
\title{Create a \code{CScaler} object to conditionally scale}
\usage{
CScaler(basis, id_vars, base_var = "base_amt", scale = 1, suffix = "scl")
}
\arguments{
\item{basis}{Data.frame of data used to perform the scale.}

\item{id_vars}{Names of columns used to join \code{data} and \code{base}.}

\item{base_var}{Name of the the base amount column is in \code{basis}.}

\item{scale}{Number used to multiply the resulting data.}

\item{suffix}{Character(1) for suffix to use in \code{.name} of \code{across}.
Default value is "scl". If no suffix is desired, set it as an empty string
\code{suffix = ""}.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create a \code{CScaler} object to conditionally scale.
}
\details{
Create a \code{CScaler} object used to conditionally rescale to a maximum
amount conditional on some criteria. A typical application, for example, is
to scale all expenses based on the sales per year.
}
\examples{
basis <- data.frame(
  year = c(2020, 2021),
  amt = c(1250000, 1500000)
)
cscaler <- CScaler(
  basis = basis, id_vars = "year",
  base_var = "amt", scale = 1000
)
stopifnot(S7::S7_inherits(cscaler, class = CScaler))
}
