% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data2clean}
\alias{data2clean}
\title{Test Data Set to Clean}
\format{
Data frame with 30 rows and 29 columns.
\describe{
\item{ndx}{TBA}
\item{not_uniq}{TBA}
\item{ran_pct}{TBA}
\item{ran_int}{TBA}
\item{ran_color}{TBA}
\item{ran_dat}{TBA}
\item{is_rare}{TBA}
\item{bool10}{TBA}
\item{bool25}{TBA}
\item{bool50}{TBA}
\item{same_num}{TBA}
\item{nzv_num}{TBA}
\item{int_num}{TBA}
\item{clean_int_num}{TBA}
\item{inth_num}{TBA}
\item{normal_num}{TBA}
\item{abnormal_num}{TBA}
\item{lognormal_num}{TBA}
\item{logabnormal_num}{TBA}
\item{log1ps10}{TBA}
\item{err_int}{TBA}
\item{same_text}{TBA}
\item{pos_text}{TBA}
\item{cont_text}{TBA}
\item{len_cont_text}{TBA}
\item{clean_text}{TBA}
\item{len_clean_text}{TBA}
\item{cont_date}{TBA}
\item{clean_date}{TBA}
}
}
\usage{
data2clean
}
\description{
Test data set to clean.
}
\details{
An example of data set used to test the cleaning functions.
}
\keyword{datasets}
