% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data2clean_ddict}
\alias{data2clean_ddict}
\title{Test Data Dictionary}
\format{
Data frame with 29 rows and 13 columns.
\describe{
\item{table}{TBA}
\item{raw_name}{TBA}
\item{name}{TBA}
\item{label}{TBA}
\item{raw_dtype}{TBA}
\item{dtype}{TBA}
\item{role}{TBA}
\item{process}{TBA}
\item{rule}{TBA}
\item{desc}{TBA}
\item{note}{TBA}
\item{name1}{TBA}
\item{dtype1}{TBA}
}
}
\usage{
data2clean_ddict
}
\description{
Test data dictionary.
}
\details{
An example of data dictionary  based on \code{data2clean}.
}
\keyword{datasets}
