% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{DDict}
\alias{DDict}
\title{Create a Data Dictionary}
\usage{
DDict(data = NULL)
}
\arguments{
\item{data}{Data dictionary.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create a data Dictionary.
}
\details{
Create data Dictionary with the following properties.
\describe{
\item{data}{Dataframe of info on the variables. See details.}
\item{data_path}{Path where files are saved.}
\item{data_base_fn}{Base file name to export data. Default is "ddict_data".}
\item{status_base_fn}{Base file name to export status, Default is "ddict_status".}
\item{data_fn}{File name to export data.
\code{file.path(data_path, data_base_fn, Sys.Date())}}
\item{status_fn}{Base file name to export status, Default is "ddict_status".
\code{file.path(data_path, status_base_fn, Sys.Date())}}
\item{dtypes}{Character vector of data types allowed by \code{DDict}.
Read-only property.}
}
}
\section{Content of \code{data}}{


The \code{data} property is a data.frame with:
\describe{
\item{table}{Name of the data.frame containing the variable.}
\item{raw_name}{Name of the variable as shown in the original data.frame.}
\item{name}{New name to assign to the variable.}
\item{label}{Label to identify the variable, e.g. used by \pkg{labelled}.}
\item{raw_dtype}{Data type of raw data.}
\item{dtype}{Data type of used data.}
\item{role}{Text to identify the role of a variable.
No validity check performed. e.g. "key" to flag keys in the data.}
\item{process}{Text to identify transformations, etc., applied to the
variable. No validity check performed. For example, "lg" could be used to
flag a variable for logarithmic transformation.}
\item{rule}{Text to identify a rule applicable to the variable. Typically
used with the \pkg{validate} package. No validity check performed.}
\item{desc}{Description.}
\item{note}{Some notes.}
}
}

\examples{
ddict <- DDict()
stopifnot(S7::S7_inherits(ddict, class = DDict))
}
