% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_cast.R
\name{ddict_cast}
\alias{ddict_cast}
\title{Cast data types of columns using a \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data.frame with variables to cast to a data type.}

\item{is_raw_nm}{\code{FALSE} = use the \code{name} from
\code{DDict}; \code{TRUE} = use \code{raw_name} from \code{DDict}. Default
is \code{FALSE}.}

\item{table_nm}{Name of table. Used when doing loop or when \code{data}
is from a function argument.}
}
\value{
\code{data} with new data types.
}
\description{
Cast data types of columns using a \code{DDict}.
}
\details{
The columns data type is recast to the data type specified in \code{dtype}
when it differ from the data type specified in \code{raw_dtype}. Nothing is
done When \code{dtype} is \code{NA} or empty and a warning is issued.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\code{\link[=cast_data]{cast_data()}}
}
