% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_filter.R
\name{ddict_filter}
\alias{ddict_filter}
\title{Filter \code{data} from a \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{table_nm}{Table name used to filter the dictionary. See details for
more info.}

\item{role_rgx}{Regular expression to filter \strong{role}. See details for more
info.}

\item{process_rgx}{Regular expression to filter \strong{process}. See details for
more info.}

\item{rule_rgx}{Regular expression to filter \strong{rule}. See details for more
info.}

\item{pull_var}{String to identify variable to pull with \code{dplyr::pull}.
If \code{NULL}, the entire table is returned. Default value is \code{NULL}.}
}
\value{
Filtered \code{data} from \code{DDict} object.
}
\description{
Filter \code{data} from a \code{DDict}.
}
\details{
The records are filtered using regular expressions. If no criteria is
provided for \strong{table_nm}, \strong{role}, \strong{process} and \strong{rule}, the full
data is returned. When the filter is \code{NA}, rows with \code{NA} are
returned.
}
\examples{
\dontrun{
TODO
}
}
