% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_xl.R
\name{ddict_read_xl}
\alias{ddict_read_xl}
\title{Create \code{DDict} Object From Data in Excel}
\usage{
ddict_read_xl(path, file = "ddict.xlsx", sheet = "data")
}
\arguments{
\item{path}{Path where the excel is located.}

\item{file}{Name of the excel file. Default is \emph{ddict.xlsx}.}

\item{sheet}{Name of Excel sheet. Default is \emph{data}.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create \code{DDict} object from data in Excel.
}
\details{
Read an Excel file and load it into an object of class \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
