% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_ren.R
\name{ddict_ren}
\alias{ddict_ren}
\title{Rename columns using a \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data.frame with variables to rename.}

\item{table_nm}{Name of table. Used when doing loop or when \code{data}
is from a function argument.}
}
\value{
\code{data} with renamed columns.
}
\description{
Rename columns using a \code{DDict}.
}
\details{
The information is stored in an object of class \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
