% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_status.R
\name{ddict_status}
\alias{ddict_status}
\title{Analyse the status of a table in \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data.frame with variables to analyse.}

\item{table_nm}{Name of the table.}

\item{do_abort}{\code{TRUE}: Throw error message when there
is outstanding status. Default is \code{TRUE}.}
}
\value{
Data frame with status information.
}
\description{
Analyse the status of a table in \code{DDict}.
}
\details{
A data frame is returned with the following information
\describe{
\item{table}{Name of the table.}
\item{variable}{Name of the variable.}
\item{is_ddict}{Flag. \code{TRUE}: the variable is in the data dictionary,
\code{FALSE} if it is not.}
\item{is_data}{Flag. \code{TRUE}: the variable is in the data,
\code{FALSE} if it is not.}
}
}
\examples{
\dontrun{
TODO
}
}
