% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{ddict_table}
\alias{ddict_table}
\title{Data about a table from a \code{DDict}}
\usage{
ddict_table(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{table_nm}{Name of the table.}
}}
}
\value{
\code{data} from \code{DDict} object.
}
\description{
Data about a table from a \code{DDict}.
}
\details{
An error message is returned if no data on the table is available.
}
\examples{
\dontrun{
TODO
}
}
