% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_transf_dat.R
\name{ddict_transf_dat_wk}
\alias{ddict_transf_dat_wk}
\title{Transform Date Variables to \code{floor(date)} by Week}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data frame with date variables.}

\item{table_nm}{Table name used to filter the dictionary.}

\item{suffix}{String use as suffix to name the new transformed variables.
Default is \emph{wk}.}
}
\value{
\code{data} with new transformed columns.
}
\description{
Transform date variables to \code{floor(date)} by week.
}
\details{
The date are converted to the first day of the week when they occur. It uses
the function \code{floor_date(x, unit = "weeks")} from \pkg{lubridate}.
}
\examples{
\dontrun{
TODO
}
}
