% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_transf_dat.R
\name{ddict_transf_dat_ym}
\alias{ddict_transf_dat_ym}
\title{Transform Date Variables to \code{yyyy.mm} Format}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data frame with date variables.}

\item{table_nm}{Table name used to filter the dictionary.}

\item{suffix}{String use as suffix to name the new transformed variables.
Default is \emph{ym}.}

\item{iso}{Logical. When \code{TRUE} the ISO year will be used for the year.
When \code{FALSE}, the calendar year will be used. Default is \code{TRUE}.}
}
\value{
\code{data} with new transformed columns.
}
\description{
Transform date variables to \code{yyyy.mm} format.
}
\details{
The variables are transformed to a decimal format where the decimal
represents the month rather than a fraction of the year as is the usual
date decimal format.
}
\examples{
\dontrun{
TODO
}
}
