% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_transf_num.R
\name{ddict_transf_num}
\alias{ddict_transf_num}
\title{Transform Numeric Variables with a Custom Function}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data frame with date variables.}

\item{fn}{Custom function.}

\item{suffix}{String use as suffix to name the new transformed variables.}

\item{table_nm}{Table name used to filter the dictionary.}
}
\value{
\code{data} with new transformed columns.
}
\description{
Transform numeric variables with a custom function.
}
\details{
The date are converted using a function defined in \code{fn}.
}
\examples{
\dontrun{
TODO
}
}
