% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_uniq.R
\name{ddict_uniq}
\alias{ddict_uniq}
\title{Validate that columns have unique values}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data.frame with variables to test for uniqueness.}

\item{table_nm}{Name of the table.}

\item{role_rgx}{Regular expression to select variables with this role.
Default value is \code{r"(\buniq\b)"}.}
}
\value{
Logical named vector with the results. TRUE is when the variable has
unique values. FALSE otherwise.
}
\description{
Validate that columns have unique values.
}
\details{
The columns are selected from the \code{DDict} when their role match
the regular expression in \code{uniq_rgx}. If \code{uniq_rgx} does not find a
match, and error will be thrown.
}
\examples{
\dontrun{
TODO
}
}
