% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_update.R
\name{ddict_update}
\alias{ddict_update}
\title{Create an Updated \code{DDict} Excel File}
\usage{
ddict_update(nms, env = .GlobalEnv)
}
\arguments{
\item{nms}{Character() with names of data.frames.}

\item{env}{Environment. Default is \code{.GlobalEnv}, don't change unless
you know what you are doing.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create an updated \code{DDict} excel file.
}
\details{
Create an xl file with extracted information on many dataframe.
}
\examples{
\dontrun{
TODO
}
}
