% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{extractDDict}
\alias{extractDDict}
\title{Extract Information About Data to a Data Dictionary}
\usage{
extractDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame from which to extract the variables' details.}
\item{table_nm}{Name of table. Used when doing loop or when \code{data}
is from a function argument.}
}}
}
\value{
Object of class \code{DDict}.
}
\description{
Extract information about data to a data dictionary.
}
\details{
The information will be stored in an object of class \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
