% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{filterDDict}
\alias{filterDDict}
\title{Filter from a \code{DDict}}
\usage{
filterDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{table_rgx}{Regular expression to filter \strong{table}.}
\item{vtype_rgx}{Regular expression to filter \strong{vtype}.}
\item{process_rgx}{Regular expression to filter \strong{process}.}
\item{rule_rgx}{Regular expression to filter \strong{rule}.}
}}
}
\value{
\code{data} from \code{DDict} object.
}
\description{
Filter from a \code{DDict}.
}
\details{
The records are filtered using regular expressions.
}
\examples{
\dontrun{
TODO
}
}
