% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flrpkgPreproc-package.R
\docType{package}
\name{flrpkgPreproc}
\alias{flrpkgPreproc-package}
\alias{flrpkgPreproc}
\title{Preprocessing Functions and Objects for Ephel}
\description{
Preprocessing functions and objects for Ephel.
}
\details{
Functions and objects used for preprocessing data by Ephel.
}
\section{flrpkgPreproc functions}{

The functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FrankLef/flrpkgPreproc}
  \item Report bugs at \url{https://github.com/FrankLef/flrpkgPreproc/issues}
}

}
\author{
\strong{Maintainer}: François Lefebvre \email{flefebvre01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
