% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules.R
\name{hbRules}
\alias{hbRules}
\title{Get the Range Using a Factor of the Median}
\usage{
hbRules(object, ...)
}
\arguments{
\item{object}{Object of class \code{Rules}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame to process.}
\item{cols}{Character vector of column names to process.}
\item{tol}{Tolerance number. Must be between 1 and 100. Default value
is 80/20.}
\item{na.rm}{Flag used by \code{median(x, na.rm)}. Default value is
\code{TRUE}.}
}}
}
\value{
List of ranges.
}
\description{
Get the range using a factor of the median.
}
\details{
This will compute ranges where the lower boundary is \code{median / tol} and
the upper boundary is \code{median * tol}. The tolerance \code{tol} is
loosely (very loosely) based on hidiroglou-berthelot which is why the
function is called \code{hbRules}.

info:
Normally, multiplying then dividing the median to obtain boundaries is more
appropriate for skewed data. In a business case, the data is very often
skewed which is fine in that case. However in a business context, when
applied to more centrally distributed data,  such as as log-transformed sales
amount per invoice for example, it is unlikely that high
amounts are mistakes and should be selected. However, very large amounts
should. In such a practical business context, experience has shown this
method to be quite relevant.
}
\examples{
\dontrun{
TODO
}
}
