% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{labelDDict}
\alias{labelDDict}
\title{Add Labels to Columns Using a \code{DDict}.}
\usage{
labelDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame with variables to label.}
\item{table_nm}{Name of the table.}
}}
}
\value{
\code{data} with labels added to the columns.
}
\description{
Add labels to columns using a \code{DDict}.
}
\details{
The labels, stored in an object of class \code{DDict}, are used by
\pkg{sjlabelled} to add labels. If the label is empty or \code{NA}, no label
is assigned.
}
\examples{
\dontrun{
TODO
}
}
