% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalz.R
\name{Normalz}
\alias{Normalz}
\title{Create a Normalz Object}
\usage{
Normalz(basis, id_vars, base_var = "base_amt", scale = 1, sufx = "nmz")
}
\arguments{
\item{basis}{Data.frame of data used to perform the normalization.}

\item{id_vars}{Names of columns used to join \code{data} and \code{base}.}

\item{base_var}{Name of the the base amount column is in \code{basis}.}

\item{scale}{Number used to multiply the resulting data.}

\item{sufx}{Character(1) for suffix to use in \code{.name} of \code{across}.
Default value is "nmz". If no suffix is desired, set it as an empty string
\code{sufx = ""}.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create a Normalz object.
}
\details{
Create a \code{Normalz} object used to normalize conditionally with a given
amount. A typical application, for example, is to normalize all expense based
on the sales per year.
}
\examples{
basis <- data.frame(
  year = c(2020, 2021),
  amt = c(1250000, 1500000)
)
normlz <- Normalz(
  basis = basis, id_vars = "year",
  base_var = "amt", scale = 1000
)
stopifnot(S7::S7_inherits(normlz, class = Normalz))
}
