% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalz.R
\name{normalz_do}
\alias{normalz_do}
\title{Apply conditional normalization}
\usage{
normalz_do(object, ...)
}
\arguments{
\item{object}{Object of class \code{Normalz}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame to normalize.}
\item{vars}{Columns to normalize.}
\item{inverse}{Perform inverse normalization. The suffix \code{suffix} is
not used when \code{inverse=TRUE}}
\item{keep}{Keep the \code{base_var} in the output.}
}}
}
\value{
Data.frame.
}
\description{
Apply conditional Normalization.
}
\details{
Normalize \code{data} depending on the \code{id_vars} from \code{Normalz}.
}
\examples{
\dontrun{
TODO
}
}
