% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_reset}
\alias{prune_reset}
\title{Reset the Prune ID Column}
\usage{
prune_reset(data, id_var = "prune_id", default = "ok")
}
\arguments{
\item{data}{Data.frame.}

\item{id_var}{Name of the column with prune ID. Default is \emph{prune_id}.}

\item{default}{Default value in the id_var column. Default is \emph{ok}.}
}
\value{
Data.frame with new/reset id_var column.
}
\description{
Reset the prune ID column.
}
\details{
The column will be created if it doesn't exist. The column will be entirely
overwritten with \code{default}.
}
\examples{
df <- data.frame(x = letters[1:3])
df <- prune_reset(df)
stopifnot(
  identical(df, data.frame(x = letters[1:3], prune_id = "ok"))
)

}
