% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_upd}
\alias{prune_upd}
\title{Update the Prune ID Column.}
\usage{
prune_upd(data, flags, id, id_var = "prune_id", default = "ok")
}
\arguments{
\item{data}{Data.frame.}

\item{flags}{Logical vector.}

\item{id}{String value to update to. Must NOT be the same as \code{default}.}

\item{id_var}{Name of the prune id column. Default is \emph{prune_id}.}

\item{default}{Default value in the \code{id_var} column. Default is \emph{ok}.}
}
\value{
Data.frame with updated \code{id_var} column.
}
\description{
Update the prune ID column.
}
\details{
The \code{id_var} column will be udated with the \code{id} value whenever
the corresponding element in \code{flags} will be \code{TRUE} \bold{and}
the value in \code{id_var} will be equal to \code{default}.
}
\examples{
df <- data.frame(x = letters[1:3], prune_id = c("ok", "ok", "X"))
df <- prune_upd(df, flags = c(TRUE, FALSE, TRUE), id = "Y")
stopifnot(
  identical(df, data.frame(x = letters[1:3], prune_id = c("Y", "ok", "X")))
)
}
\seealso{
\code{\link[=prune_reset]{prune_reset()}}
}
