% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{renDDict}
\alias{renDDict}
\title{Rename Columns Using a \code{DDict}}
\usage{
renDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame with variables to rename.}
\item{table_nm}{Name of the table.}
}}
}
\value{
\code{data} with renamed columns.
}
\description{
Rename columns using a \code{DDict}.
}
\details{
The information is stored in an object of class \code{DDict}.
}
\examples{
\dontrun{
TODO
}
}
