% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{rmDDict}
\alias{rmDDict}
\title{Remove a Variable from a Data Dictionnary}
\usage{
rmDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{table}{Name of table from which to remove the variable.}
\item{raw_name}{Name of variable to remove.}
}}
}
\value{
Object of class \code{DDict}
}
\description{
Remove a variable from a data dictionary.
}
\details{
Remove all information about a variable from a DDictionary.
}
\examples{
\dontrun{
TODO
}
}
