% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules.R
\name{Rules}
\alias{Rules}
\title{Create a \code{Rules} Object}
\usage{
Rules()
}
\value{
Object of class \code{Rules}.
}
\description{
Create a \code{Rules} object.
}
\details{
Create a \code{Rules} object used to to managed validation rules as they are
typically used by the \pkg{validate} package.
}
\examples{
\dontrun{
TODO
}
}
