% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{statusDDict}
\alias{statusDDict}
\title{Analyse the Status of a Table in \code{DDict}}
\usage{
statusDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame with variables to label.}
\item{do_abort}{TRUE (default): Throw error message when there
is outstanding status.}
\item{table_nm}{Name of the table.}
}}
}
\value{
Data frame with status information.
}
\description{
Analyse the status of a table in \code{DDict}.
}
\details{
A data frame is returned with the following information
\describe{
\item{table}{Name of the table.}
\item{variable}{Name of the variable.}
\item{is_ddict}{Flag. \code{TRUE}: the variable is in the data dictionary,
\code{FALSE} if it is not.}
\item{is_data}{Flag. \code{TRUE}: the variable is in the data,
\code{FALSE} if it is not.}
}
}
\examples{
\dontrun{
TODO
}
}
