% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDict.R
\name{TDict}
\alias{TDict}
\title{Create a Table Dictionary}
\usage{
TDict(data = NULL)
}
\arguments{
\item{data}{Table dictionary.}
}
\value{
Object of class \code{TDict}.
}
\description{
Create a table dictionary.
}
\details{
Create table Dictionary with the following properties.
\describe{
\item{data}{Dataframe of info on the variables. See details.}
\item{data_path}{Path where the dictionary is saved, set to \code{getwd()}.}
\item{data_base_fn}{Base file name to export data. Default is \emph{tdict_raw}.}
\item{ext}{Extension fo as a suffix to file names. Default is \emph{"xlsx"}.}
\item{data_fn}{Full file name to export the \emph{tdict} data.}
}
}
\section{Content of \code{data}}{


The \code{data} property is a data.frame with:
\describe{
\item{path}{Directory of data source.}
\item{file}{File name of data source.}
\item{table}{Name of query/table (database) or spreadsheet (excel).}
\item{name}{Name of table.}
\item{raw_name}{Import name of table. Usually begins with \emph{raw_"}.}
\item{label}{Label for the table.}
\item{type}{Source type of raw data. e.g. \emph{accdb}, \emph{xlsx}, etc.}
\item{role}{String to identify the role of the table.}
\item{process}{Text to identify how to process the table}
\item{rule}{Text to identify a set of rules applicable to the table.}
\item{desc}{Description.}
\item{note}{Some notes.}
}
}

\examples{
tdict <- TDict()
stopifnot(S7::S7_inherits(tdict, class = TDict))
}
