% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdict_load_accdb.R
\name{tdict_load_accdb}
\alias{tdict_load_accdb}
\title{Extract a Data Set from MS Access}
\usage{
tdict_load_accdb(path, file, qry)
}
\arguments{
\item{path}{Path where MS Access database is located.}

\item{file}{Name of MS Access database.}

\item{qry}{Name of table/query.}
}
\value{
Data.frame.
}
\description{
Extract a data set from MS Access.
}
\details{
Extract data from MS Access. Regardless of whether it is a query or a table.
}
\examples{
\dontrun{
TODO
}
}
