% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdict_load_accdb.R
\name{tdict_load_accdb_many}
\alias{tdict_load_accdb_many}
\title{Extract Many Data Sets from MS Access}
\usage{
tdict_load_accdb_many(
  object,
  raw_nm = TRUE,
  type = "accdb",
  process = "load",
  env = .GlobalEnv
)
}
\arguments{
\item{object}{Obkject of class \code{TDict}.}

\item{raw_nm}{Flag, TRUE = use raw name to name the data.frame. Otherwise
use the regular name. Default is TRUE.}

\item{type}{String to filter the type in \code{TDict}. Default is \emph{accdb}.}

\item{process}{String to filter the process in \code{TDict}. Default is \emph{load}.}

\item{env}{Environement used to create the data.frame. Don't change this
unless you know what you are doing.}
}
\value{
Count of dataframe extracted.
}
\description{
Extract many data sets from MS Access.
}
\details{
The extraction is done using the information from an object of class
\code{TDict}.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\code{\link[=tdict_load_accdb]{tdict_load_accdb()}}
}
