% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdict_read_xl.R
\name{tdict_read_xl}
\alias{tdict_read_xl}
\title{Create \code{TDict} Object From Data in Excel}
\usage{
tdict_read_xl(path, file = "tdict.xlsx", sheet = "data")
}
\arguments{
\item{path}{Path where the excel is located.}

\item{file}{Name of the excel file. Default is \emph{tdict.xlsx}.}

\item{sheet}{Name of Excel sheet. Default is \emph{data}.}
}
\value{
Object of class \code{TDict}.
}
\description{
Create \code{TDict} object from data in Excel.
}
\details{
Read an Excel file and load it into an object of class \code{TDict}.
}
\examples{
\dontrun{
TODO
}
}
