% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdict_xl.R
\name{tdict_read_xl_trim}
\alias{tdict_read_xl_trim}
\title{Create \code{TDict} Object From Filtered Data in Excel}
\usage{
tdict_read_xl_trim(path, file = "tdict.xlsx", sheet = "data", process = "load")
}
\arguments{
\item{path}{Path where the excel is located.}

\item{file}{Name of the excel file. Default is \emph{tdict.xlsx}.}

\item{sheet}{Name of Excel sheet. Default is \emph{data}.}

\item{process}{String used to filter the \code{process} column. Default
to \emph{load}.}
}
\value{
Object of class \code{TDict}.
}
\description{
Create \code{TDict} object from filtered data in Excel.
}
\details{
Read an Excel file and load it into an object of class \code{TDict}. The data
is filtered on column \code{process} with \emph{load}.
}
\examples{
\dontrun{
TODO
}
}
