% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_dat_dur.R
\name{transf_dat_dur}
\alias{transf_dat_dur}
\title{Compute Duration in Days for a Pairs of Columns.}
\usage{
transf_dat_dur(data, dur_var, start_var, end_var, digits = 0)
}
\arguments{
\item{data}{Data frame to process.}

\item{dur_var}{String. Name of duration variable.}

\item{start_var}{String. Name of start variable.}

\item{end_var}{String. Name of end variable.}

\item{digits}{Integer for digits to round the days.}
}
\value{
Data frame with new column of duration.
}
\description{
Compute duration in days for a pairs of columns.
}
\details{
Compute the duration in the new column \code{dur} using start date
from \code{start} and end date using \code{end}. This function is used by
\code{transf_dat_dur_many}.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\code{\link[=transf_dat_dur_many]{transf_dat_dur_many()}}
}
