% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_dat_dur.R
\name{transf_dat_dur_many}
\alias{transf_dat_dur_many}
\title{Compute Duration in Days for Several Pairs of Columns.}
\usage{
transf_dat_dur_many(data, vars, table_nm = deparse1(substitute(data)))
}
\arguments{
\item{data}{Data frame to process.}

\item{vars}{Data frame of variables with the following columns
\describe{
\item{dur}{Name of the new column with the duration.}
\item{start}{Name of the column with start dates.}
\item{end}{Name of the column with end dates.}
}}

\item{table_nm}{Table name. Used for error message.}
}
\value{
Data frame with new duration columns in days.
}
\description{
Compute duration in days for several pairs of columns.
}
\details{
Compute the duration in the new column \emph{dur} using start date
from \emph{start} and end date using \emph{end}.
}
\examples{
\dontrun{
TODO
}
}
