% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_cbind.R, R/valid_tools.R
\name{valid_cbind}
\alias{valid_cbind}
\title{Add the Flags from \code{validation} to Data}
\usage{
valid_cbind(data, validation, suffix)

valid_cbind(data, validation, suffix)
}
\arguments{
\item{data}{Data.frame.}

\item{validation}{Object of class \code{validation} from \pkg{validate}.}

\item{suffix}{String to add as a suffix to the new columns.}
}
\value{
Augmrnted data.frame with new columns.

Augmented data.frame with new columns.
}
\description{
Add the Flags from \code{validation} to data.

Add the Flags from \code{validation} to data.
}
\details{
Add the results of an object of class \code{validation} fro \pkg{validate}
To the data. The new columns will have a \code{suffix}. One last logical
column called \code{paste("is", suffix, sep = "_")} is added with value
\code{TRUE} is any of the column with the \code{suffix} are \code{TRUE}.

Add the results of an object of class \code{validation} fro \pkg{validate}
To the data. The new columns will have a \code{suffix}. One last logical
column called \code{paste("is", suffix, sep = "_")} is added with value
\code{TRUE} is any of the column with the \code{suffix} are \code{TRUE}.
}
\examples{
\dontrun{
TODO
}
\dontrun{
TODO
}
}
