% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_msg.R
\name{valid_msg_pct}
\alias{valid_msg_pct}
\title{Write information message on valid/invalid pct}
\usage{
valid_msg_pct(validation, suffix)
}
\arguments{
\item{validation}{Object of class \code{validation} created with
\code{validate::validator}.}

\item{suffix}{Suffix representing the type of validation.}
}
\value{
String used for message.
}
\description{
Write information message on valid/invalid pct.
}
\details{
Write information message on valid/invalid pct. Mainly use with \pkg{logging}.
}
\examples{
\dontrun{
TODO
}
}
