% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_range.R
\name{valid_range_madn}
\alias{valid_range_madn}
\title{Get the range using the normalized MAD}
\usage{
valid_range_madn(data, cols, tol = 0.95, na.rm = TRUE)
}
\arguments{
\item{data}{Data.frame .}

\item{cols}{Character() with the names of the columns to compute.}

\item{tol}{A percentage representing the probabilities, or proportion of data,
to keep. It is the \code{p} in \code{stats::qchisq(prob, df = 1)}. Default
is 0.95. Must be between 0.51 and 0.9999.}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\value{
List of ranges with one element per column.
}
\description{
Get the range using the normalized MAD.
}
\details{
This method is applicable when the data is more or less centralized.
It is robust to deviations. It uses the \emph{median absolute deviation, MAD}.
It is described in many books.  For example, it could also be computed using
the \pkg{performance} package with \code{performance::check_outliers()}.

\code{tol} is a percentage representing the probabilities, or proportion of data,
to keep. It is the \code{p} in \code{stats::qchisq(prob, df = 1)}. The
outliers are selected from the left and right tails. That is, for example,
with \code{tol = 0.95} 2.5\% will be selected from the left and 2.5\% will be
selected from the right.
}
\section{\pkg{Performance}}{

If more outlier detection methods are required, please see the
\pkg{performance} at
\href{https://easystats.github.io/performance/index.html}{performance}. This
package is packed with very useful utilities.
}

\examples{
\dontrun{
TODO
}
}
