% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_rules.R
\name{valid_rules}
\alias{valid_rules}
\title{Create data.frame of out-of-bound rules}
\usage{
valid_rules(ranges, suffix = c("oob", "madn"))
}
\arguments{
\item{ranges}{List of variable names with their respective range.}

\item{suffix}{Suffix to add to create the rules name. Default to 'oob'.}
}
\value{
Data frame of rules usable by \code{validate::validator()}.
}
\description{
Create data.frame of out-of-bound rules used by \pkg{validate}.
}
\details{
The data.frame includes the name of the rule, its label and the rule itself.
All rule names end with the \code{suffx} '_oob'.
}
\examples{
\dontrun{
TODO
}
}
