% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{cast_data}
\alias{cast_data}
\title{Cast data to new data Type}
\usage{
cast_data(object, data, var, dtype, table_nm)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data frame.}

\item{var}{Name of the variable to cast.}

\item{dtype}{Name of data type.}

\item{table_nm}{Name of table. Only used by error message.}
}
\value{
\code{data} with new data type.
}
\description{
Cast data to new data type.
}
\details{
This function is used by \code{ddict_cast()}. When \code{dtype} is invalid,
an error message is issued.
}
\seealso{
ddict_cast
}
