% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{cast_data}
\alias{cast_data}
\title{Cast Data to New Data Type}
\usage{
cast_data(object, x, dtype)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{x}{Vector.}

\item{dtype}{Name of data type.}
}
\value{
x with new data type.
}
\description{
Cast data to new data type.
}
\details{
This function is used by \code{castDDict()}. When \code{dtype} is invalid,
an error message is issued.
}
\seealso{
\code{\link[=castDDict]{castDDict()}}
}
