% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{castDDict}
\alias{castDDict}
\title{Cast Data Types of Columns Using a \code{DDict}.}
\usage{
castDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{data}{Data.frame with variables to label.}
\item{table_nm}{Name of the table.}
\item{is_raw_nm}{\code{FALSE} (default) = use the \code{name} from
\code{DDict}; \code{TRUE} = use \code{raw_name} from \code{DDict}.}
}}
}
\value{
\code{data} with recasted columns.
}
\description{
Cast data types of columns using a \code{DDict}.
}
\details{
The columns data type is recast to the data type specified in \code{dtype}
when it differ from the data type specified in \code{raw_dtype}. Nothing is
done When \code{dtype} is \code{NA} or empty.
}
\examples{
\dontrun{
TODO
}
}
