% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_NA}
\alias{clean_NA}
\title{Replace empty string and \code{NaN, -Inf, Inf} with \code{NA}}
\usage{
clean_NA(
  data,
  na_patterns = c("^[[:punct:]]$", "_NA", "^[Nn][/\\\\\\\\[[:space:]]]*[Aa]$")
)
}
\arguments{
\item{data}{Data frame.}

\item{na_patterns}{Character vector of regular expressions to run with
\code{stringr::str_replace(x, pattern = stringr::regex(pattern, ignore_case = TRUE),
replacement = NA_character_)}.}
}
\value{
Data frame.
}
\description{
Replace empty string and \code{NaN, -Inf, Inf} with \code{NA}.
}
\details{
This function calls \code{clean_inf} and \code{clean_empty}.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\code{\link[=clean_inf]{clean_inf()}} \code{\link[=clean_empty]{clean_empty()}}
}
