% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data2clean}
\alias{data2clean}
\title{Test Data Set to Clean}
\format{
Data frame with 30 rows and 29 columns.
\describe{
\item{ndx}{Integer. Index.}
\item{not_uniq}{Integer. One of the number is not unique.}
\item{ran_pct}{Double. Random percentages / probabilities.}
\item{ran_int}{Integer. Random integers.}
\item{ran_color}{Character. Random primary, secondary or tertiary color names.}
\item{ran_dat}{Random dates.}
\item{is_rare}{Logical. Only one \code{TRUE}.}
\item{bool10}{Logical. 10\% of \code{TRUE}.}
\item{bool25}{Logical. 25\% of \code{TRUE}.}
\item{bool50}{Logical. 50\% of \code{TRUE}.}
\item{same_num}{Numerical. Same number everywhere.}
\item{nzv_num}{Numerical. Near-Zero-Variance data.}
\item{int_num}{TBA}
\item{clean_int_num}{TBA}
\item{inth_num}{TBA}
\item{normal_num}{TBA}
\item{abnormal_num}{TBA}
\item{lognormal_num}{TBA}
\item{logabnormal_num}{TBA}
\item{log1ps10}{TBA}
\item{err_int}{TBA}
\item{same_text}{TBA}
\item{pos_text}{TBA}
\item{cont_text}{TBA}
\item{len_cont_text}{TBA}
\item{clean_text}{TBA}
\item{len_clean_text}{TBA}
\item{cont_date}{TBA}
\item{clean_date}{TBA}
}
}
\usage{
data2clean
}
\description{
Test data set to clean.
}
\details{
An example of data set used to test the cleaning functions.
}
\keyword{datasets}
