% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data2clean_ddict}
\alias{data2clean_ddict}
\title{Test Data Dictionary}
\format{
Data frame with 29 rows and 13 columns.
\describe{
\item{table}{Name of the data frame containing the variable.}
\item{raw_name}{Name of the variable as shown in the original data frame.}
\item{name}{New name to assign to the variable.}
\item{label}{Label to use with the variable, e.g. used by \pkg{labelled}.}
\item{raw_dtype}{Data type of raw data.}
\item{dtype}{Data type of used data.}
\item{role}{Text to identify the role of a variable.}
\item{process}{Text to identify processes associated used.}
\item{rule}{Text to identify the rules applicable to the variable}
\item{desc}{Description.}
\item{note}{Note.}
\item{name1}{Alternative \code{name}.}
\item{dtype1}{Alternative \code{dtype}.}
}
}
\usage{
data2clean_ddict
}
\description{
Test data dictionary.
}
\details{
An example of data dictionary  based on \code{data2clean}.
}
\keyword{datasets}
