% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{DDict}
\alias{DDict}
\title{Create a Data Dictionary}
\usage{
DDict(data = NULL)
}
\arguments{
\item{data}{Data dictionary.}
}
\value{
Object of class \code{DDict}.
}
\description{
Create a data Dictionary.
}
\details{
Create data Dictionary with the following properties.
\describe{
\item{data}{Dataframe of info on the variables. See details.}
\item{dtypes}{Character vector of data types allowed by \code{DDict}.
Read-only property.}
}
}
\section{Content of \code{data}}{


The \code{data} property is a data.frame with:
\describe{
\item{table}{Name of the data.frame containing the variable.}
\item{raw_name}{Name of the variable as shown in the original data.frame.}
\item{name}{New name to assign to the variable.}
\item{label}{Label to identify the variable, e.g. used by \pkg{labelled}.}
\item{raw_dtype}{Data type of raw data.}
\item{dtype}{Data type of used data.}
\item{vtype}{Variable type. Discretionary type assigned to the variable.
No validity check performed. e.g. "key" to flag keys in the data.}
\item{desc}{Optional description.}
\item{note}{Optional note.}
}
}

\examples{
ddict <- DDict()
stopifnot(S7::S7_inherits(ddict, class = DDict))
}
