% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{DDict}
\alias{DDict}
\title{Create a data Dictionary}
\usage{
DDict(data = class_missing)
}
\arguments{
\item{data}{Data.frame of info on the variables.}
}
\value{
Object of class DDict.
}
\description{
Create a data Dictionary.
}
\details{
Create data Dictionary with the following properties.
\describe{
\item{data}{Dataframe of info on the variables. See details.}
}
}
\section{Content of \code{data}}{


The \code{data} property is a data.frame with:
\describe{
\item{table}{Name of the data.frame containing the variable.}
\item{raw_name}{Name of the variable as shown in the original data.frame.}
\item{name}{New name to assign to the variable..}
\item{label}{Label to identify the variable, e.g. used by \pkg{sjlabelled}.}
\item{desc}{Description of the variable.}
\item{note}{Discretionary note on the variable.}
\item{raw_dtype}{Data type of raw data.}
\item{dtype}{Data type of used data.}
}
}

\examples{
DDict <- DDict()
}
