% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{ddict_filter}
\alias{ddict_filter}
\title{Filter from a \code{DDict}}
\usage{
ddict_filter(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{table_nm}{Compulsory name of the table..}
\item{role_rgx}{Regular expression to filter \strong{role}.}
\item{process_rgx}{Regular expression to filter \strong{process}.}
\item{rule_rgx}{Regular expression to filter \strong{rule}.}
}}
}
\value{
\code{data} from \code{DDict} object.
}
\description{
Filter from a \code{DDict}.
}
\details{
The records are filtered using regular expressions. If no criteria is
provided for \strong{role}, \strong{process} and \strong{rule}, the full data is returned.
}
\examples{
\dontrun{
TODO
}
}
