% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddict_label.R
\name{ddict_label}
\alias{ddict_label}
\title{Set labels to columns using a \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{data}{Data.frame with variables to label.}

\item{is_raw_nm}{\code{FALSE} = use the \code{name} from
\code{DDict}; \code{TRUE} = use \code{raw_name} from \code{DDict}.
Default is \code{FALSE}.}

\item{table_nm}{Table name used to filter the dictionary.}
}
\value{
\code{data} with labels added to the columns.
}
\description{
Set labels to columns using a \code{DDict}.
}
\details{
The labels, stored in an object of class \code{DDict}, are used by
to set labels with \code{labelled::var_label()}. If the label in
\code{DDict} is empty or \code{NA}, an error is returned.
}
\examples{
\dontrun{
TODO
}
}
