% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{ddict_table}
\alias{ddict_table}
\title{Data About a Table from a \code{DDict}}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{table_nm}{Table name used to filter the dictionary. If \code{NULL}
then all the data for all tables is returned. Default value is \code{NULL}.}
}
\value{
\code{data} from \code{DDict} object.
}
\description{
Data about a table from a \code{DDict}.
}
\details{
An error message is returned if no data on the table is available.
}
\examples{
\dontrun{
TODO
}
}
