% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDict.R
\name{tableDDict}
\alias{tableDDict}
\title{Data About a Table from a \code{DDict}}
\usage{
tableDDict(object, ...)
}
\arguments{
\item{object}{Object of class \code{DDict}.}

\item{...}{Additional arguments used by methods. Such as
\describe{
\item{table}{Regular expression to select the table. Default is \code{".*"}.}
}}
}
\value{
\code{data} from \code{DDict} object.
}
\description{
Data about a table from a \code{DDict}.
}
\details{
The records are filtered using regular expressions (regex).
}
\examples{
\dontrun{
TODO
}
}
